Update README.md
Emurasoft EmEditor Pro v25.0.1 x64 Multilingual

EmEditor Text Editor is a lightweight, yet extendable, easy-to-use text editor for Windows. EmEditor is very customizable, and it supports Unicode and powerful macros. EmEditor now supports very large files - up to 248 GB or 2.1 billion lines! In Professional version there are many new features including a Large File Controller, Vertical Selection Editing, Binary Editing, Optimized Search and Replace, improved Projects Plug-in, and many more! EmEditor is now able to open even larger than 248 GB by opening a portion of the file with the new custom bar - Large File Controller. The Large File Controller allows you to specify the beginning point, end point, and range of the file to be opened. It also allows you to stop the opening of the file and monitor the real size of the file and the size of the temporary disk available.

[Download Emurasoft EmEditor Pro v25.0.1 x64 Multilingual](https://developer.team/misc-development/35402-emurasoft-emeditor-pro-v2501-x64-multilingual.html)